refactor(response): move error message to response package
