+++ src/main/java/com/dnd/wedding/global/exception/ControllerExceptionHandler.java
@@ -1,5 +1,6 @@
 package com.dnd.wedding.global.exception;
 
+import com.dnd.wedding.global.response.ErrorMessage;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
