+++ src/test/java/com/dnd/wedding/global/exception/ControllerExceptionHandlerTest.java
@@ -3,6 +3,7 @@ package com.dnd.wedding.global.exception;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
+import com.dnd.wedding.global.response.ErrorMessage;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
